# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ImageToolz API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tool {
  id: string;
  name: string;
  slug: string;
  description: string;
  category: string;
  /** @nullable */
  icon?: string | null;
  usageCount: number;
  featured?: boolean;
}

export interface ToolUseInput {
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  fileSizeKb?: number | null;
}

export interface ToolUseStat {
  toolId: string;
  totalUses: number;
}

export interface StatsSummary {
  totalProcessed: number;
  totalTools: number;
  uniqueUsersToday: number;
  /** @nullable */
  topCategoryName?: string | null;
}

export interface RecentActivity {
  toolId: string;
  toolName: string;
  timestamp: string;
  /** @nullable */
  fileType?: string | null;
}

export interface ToolPopularity {
  toolId: string;
  toolName: string;
  usageCount: number;
  percentage: number;
}