# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ImageToolz API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/{id}:
    get:
      operationId: getTool
      tags: [tools]
      summary: Get a single tool
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Not found
  /tools/{id}/use:
    post:
      operationId: recordToolUse
      tags: [tools]
      summary: Record a tool usage event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolUseInput"
      responses:
        "200":
          description: Usage recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolUseStat"
...
components:
  schemas:
...
    Tool:
      type: object
      required: [id, name, slug, description, category, usageCount]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: ["string", "null"]
        usageCount:
          type: integer
        featured:
          type: boolean
    ToolUseInput:
      type: object
      properties:
        fileType:
          type: ["string", "null"]
        fileSizeKb:
          type: ["integer", "null"]
    ToolUseStat:
      type: object
      required: [toolId, totalUses]
      properties:
        toolId:
          type: string
        totalUses:
          type: integer
...