# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ImageToolz API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Tool {
  id: string;
  name: string;
  slug: string;
  description: string;
  category: string;
  /** @nullable */
  icon?: string | null;
  usageCount: number;
  featured?: boolean;
}