# lib/api-zod/src/generated/types/toolUseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ImageToolz API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ToolUseInput {
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  fileSizeKb?: number | null;
}